remove excessive test files
